Use appropriate cost-to-go computation for min possible cost in BIT*
BIT* used the cost-to-go between the start and all sampled goal states instead of the cost-to-go
between the start and the (abstract) goal. This is problematic if the goal is of type 'goal space',
because the lowest-possible cost-to-go will most likely be lower than the heuristic cost between the
start and a sampled goal. This prevented BIT* from uniformly sampling the informed set.